Optimize the code according to the CI prompt

+++ crates/bevy_render/src/texture/ktx2.rs
@@ -156,11 +156,11 @@ pub fn ktx2_buffer_to_image(
                         get_transcoded_formats(supported_compressed_formats, data_format, is_srgb);
                     let texture_format_info = texture_format;
                     let (block_width_pixels, block_height_pixels) = (
-                        texture_format_info.block_dimensions().0 as u32,
-                        texture_format_info.block_dimensions().1 as u32,
+                        texture_format_info.block_dimensions().0,
+                        texture_format_info.block_dimensions().1,
                     );
                     // Texture is not a depth or stencil format, it is possible to pass `None` and unwrap
-                    let block_bytes = texture_format_info.block_size(None).unwrap_or_default() as u32;
+                    let block_bytes = texture_format_info.block_size(None).unwrap_or_default();
 
                     let transcoder = LowLevelUastcTranscoder::new();
                     for (level, level_data) in levels.iter().enumerate() {
